+++ includes/admin/views/settings-lessons.php
@@ -1,0 +1,3 @@
+<div role="tabpanel" tabindex="0" id="lessons-panel" aria-labelledby="lessons-tab" class="box lessons">
+	<h2>Hello from lessons</h2>
+</div>