+++ includes/admin/views/settings-template.php
@@ -1,0 +1,3 @@
+<div role="tabpanel" tabindex="0" id="template-panel" aria-labelledby="template-tab" class="box template">
+	<h2>Hello from template</h2>
+</div>